# docs/swagger.yaml
basePath: /
definitions:
  dto.CreateProductInput:
    properties:
      name:
        type: string
      price:
        type: number
    type: object
  dto.CreateUserInput:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
  dto.ErrorOutput:
    properties:
      message:
        type: string
    type: object
  dto.GetJWTInput:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  dto.GetJWTOutput:
    properties:
      access_token:
        type: string
    type: object
  entity.Product:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      price:
        type: number
    type: object
host: localhost:8000
info:
  contact:
    email: nobody@gmail.com
    name: Rafael Badiale
  description: This is a simple API made for the Go Expert course.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Go Expert API Example
  version: "1.0"
paths:
  /products:
    get:
      description: Get all products data
      parameters:
      - description: page number
        in: query
        name: page
        type: string
      - description: limit
        in: query
        name: limit
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.Product'
            type: array
        "403":
          description: Forbidden
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorOutput'
      security:
      - ApiKeyAuth: []
      summary: Get all products data
      tags:
      - products
    post:
      consumes:
      - application/json
      description: Create a new product
      parameters:
      - description: product data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateProductInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorOutput'
        "403":
          description: Forbidden
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorOutput'
      security:
      - ApiKeyAuth: []
      summary: Create a new product
      tags:
      - products
  /products/{id}:
    delete:
      description: Delete a product data
      parameters:
      - description: product id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorOutput'
        "403":
          description: Forbidden
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorOutput'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorOutput'
      security:
      - ApiKeyAuth: []
      summary: Delete a product data
      tags:
      - products
    get:
      description: Get a product data
      parameters:
      - description: product id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorOutput'
        "403":
          description: Forbidden
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorOutput'
      security:
      - ApiKeyAuth: []
      summary: Get a product data
      tags:
      - products
    put:
      consumes:
      - application/json
      description: Update a product data
      parameters:
      - description: product id
        in: path
        name: id
        required: true
        type: string
      - description: product data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateProductInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorOutput'
        "403":
          description: Forbidden
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorOutput'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorOutput'
      security:
      - ApiKeyAuth: []
      summary: Update a product data
      tags:
      - products
  /user:
    post:
      consumes:
      - application/json
      description: Create user
      parameters:
      - description: user request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateUserInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorOutput'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorOutput'
      summary: Create user
      tags:
      - users
  /user/generate_token:
    post:
      consumes:
      - application/json
      description: Get a user JWT
      parameters:
      - description: user credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.GetJWTInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GetJWTOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorOutput'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorOutput'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorOutput'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorOutput'
      summary: Get a user JWT
      tags:
      - users
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"